Get things up and running again.
We've solidified the data structures here. At the level of the parser, all that
matters are the terminals and nonterminals, so we can represent these using a
simple data structure: capital strings for terminals, noncapital strings for
nonterminals; this follows Lark's conventions. The terminals' regexes, which
are necessary for SynCode, will only intervene in the lexer; the parsing logic
itself won't interact with them directly.